restructured entire login page. Placed login component on same level as app component. Now all 3 main components are positioned.
