+++ client/src/components/ChatContainerComponent.jsx
@@ -15,7 +15,10 @@ class ChatContainerComponent extends React.Component {
   render(){
     return(
       <div className='chat-container-component'>
-        <FriendsListComponent friends={this.props.friends}/>
+        <FriendsListComponent 
+          username={this.props.username}
+          deleteFriend={this.props.deleteFriend}
+          friends={this.props.friends}/>
         <ChatClientComponent 
           getFriends={this.props.getFriends}
           username={this.props.username}/>
