+++ client/src/components/FriendsListComponent.jsx
@@ -11,10 +11,17 @@ class FriendsListComponent extends React.Component {
   render(){
     return(
       <div className='friends-list-container'>
-        <h2>Friends List</h2>
+        <div className='socket-chat-title-box'>
+          <h2>Friends List</h2>
+        </div>
         <ul>
-          {this.props.friends.map((friend, i) =>
-            <li key={i}>{friend}</li>
+          {this.props.friends.map((friend, i) => 
+            <div className='friendsList'>
+                <i className="fas fa-user"></i>
+                <li 
+                  onClick={(e) => this.props.deleteFriend(this.props.username, e.target.innerHTML)}
+                  key={i}>{friend}</li>
+            </div>
           )}
         </ul>
       </div>
